// packages/core/src/llm-core/agent/types.d.ts
export type AgentAction = {
    tool: string
    toolInput: string | Record<string, any>
    toolCallId?: string
    log: string
}
export type AgentFinish = {
    returnValues: Record<string, any>
    log: string
}
export type AgentStep = {
    action: AgentAction
    observation: MessageContent
}